Tests: check camera activity in tst_qmediacapturesession
Make sure that the camera activity was correctly set when it should be.

On Android, camera activation take some time. It may happen that we are
trying to start recording on not active camera - which cause test
failure. That is why we need to wait for camera activation before test
continuation.

Activity checks perform the function of synchronization which is
necessary for android, but can be also valuable for other platforms.

Task-number: QTBUG-99098
Pick-to: 6.5
Change-Id: I65388d92fe55fb17f0aa1ca1c0fc86e84651299a
Reviewed-by: Lars Knoll <lars@knoll.priv.no>