Fix random crash with windows backend
It's not clear how to reproduce the crash, but we have crash reports.
The assumption is the crash is a side effect of used deleteLater
for a few objects in the windows backend design. In theory, objects
can have enqueued events that are handled between deleteLater() and
delete, at this time QAudioSink is already deleted.
As a result, I propose using qpointer in order to check
if the object exists.

Task-number: QTBUG-108693
Change-Id: I0c974b1faa6f0a434e95781f22984b532671daa4
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 365ef2fcc34ae3a269dd019b232e3dc9c9550f60)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>