AVFVideoSink: Avoid race condition when updating texture cache
In the main thread, in AVFVideoSinkInterface, we were updating the
texture cache by freeing it and replacing it with a newly created one.
However, in AVFVideoBuffer::textureHandle, called by the render thread,
we were accessing this texture cache to create a new texture.
This led to a race condition, where the texture cache may happen to
be invalid after being freed. To avoid, add a mutex to the
AVFVideoSinkInterface to protect access to the the texture cache.

Fixes: QTBUG-110131
Pick-to: 6.5 6.4
Change-Id: If3a4f17f51b5cb38174815b54f03961584b756b9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>