Use new permissions API on Android
[ChangeLog][Important Behavior Changes][Android] The platform
implementation no longer requests Camera and Microphone permissions,
only checks if they are granted or not. The client application must
request the permissions using the QPermission API.

Note: The WRITE_EXTERNAL_STORAGE permission is not handled by this
patch, because it's not supported by the new QPermission API for now.
This permission is still requested from the library code.

Task-number: QTBUG-109965
Change-Id: Ia51d324db3893940fe1ef18ced197cfd7b724a99
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>