Adjust multimedia examples to use new QPermission API
Now when the library itself does not request permissions, that needs
to be done on the user application level.
This patch updates all examples to query the required permissions.
As there is no QML API yet, the QML examples do it directly from
main.

Note: the WRITE_EXTERNAL_STORAGE Android permission is not handled
by this patch, because it's currently not supported by the QPermission
API.

Task-number: QTBUG-109965
Change-Id: Iaeb09f9e32fced0ecb0f5ffc63d40c64ec386cfb
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>