Rewrite screen capturing on macos
The previous generic implementation uses QScreen::grabWindow that
is not stable for both cases, screens and windows capturing.

In order to fix it, I've reimplemented it via native macos
functionality. Some part of the functionality is shared with avfcamera.

As a result, we got the following:
- fix not capturing of some windows on the screen
- enable hw acceleration for screens capturing
  (shared with camera functionality)
- enable foreign app windows capturing (by CGWindowID as WId).
  We might need some additional api for getting the list of ids.

Probably, the code is reusable for ios, to be investigated.
Screen capturing tests work locally if permissions are granted;
on CI some problems with giving the permissions.

Task-number: QTBUG-103226
Change-Id: Iada34467e4b9074eb43a1fbbdc4409e663ea668a
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit a128d5ecf40d8d670d4d3aa8b9ddc860baa3cb5e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>