AVFMediaEncoder: Set channel layout configuration when recording
When recording audio, we are setting the audio channel count to the
one specified in the encoder settings.
If this is not set, we were setting it to 2 channels by default.
However, this was causing sometimes problems as the internal
AudioConverter could not always convert from the audio capture device's
actual format to one with 2 channels. To avoid this, we should default
to setting the channel layout configuration settings based on the
channel layout of the audio device. This is safer and avoids unnecessary
format conversions.

Fixes: QTBUG-110844
Pick-to: 6.2 6.4 6.5
Change-Id: I3d78ec8b37d6780c7d96147f81ce15bd6425b448
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>