AVFMediaEncoder: Set channel count from audio device when recording
When recording audio, we are setting the audio channel count to the
one specified in the encoder settings.
If this is not set, we were setting it to 2 channels by default.
However, this was causing sometimes problems as the internal
AudioConverter could not always convert from the audio capture device's
actual format to one with 2 channels. To avoid this, we should default
to setting the channel count based on that of the audio device.
Note that in a40b9fed50ef75504cec128d4acfb1062e90df98 we set the channel
layout configuration instead of the count, but that information is not
available in 6.2.

Fixes: QTBUG-110844
Change-Id: I3d78ec8b37d6780c7d96147f81ce15bd6425b448
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit a40b9fed50ef75504cec128d4acfb1062e90df98)