Add cmake var for explicit setting of default media backend
Embedded team needs a possibility to customize the default media backend
on compile-time; gstreamer backend is likely to be used for their aims.

The most flexible solution is adding cmake env var as
it's hard to match required embedded platforms to gstreamer in QtMM.

QT_DEFAULT_MEDIA_BACKEND has been added. If it's not set,
the logic is the same as before the change.

Pick-to: 6.5
Change-Id: Id81b0e2b82edd6195c310bc8f2ad40555c6b1a2d
Reviewed-by: Lars Knoll <lars@knoll.priv.no>