Android: Integrate HW encoding for FFMpeg backend in Android
Video encoding was done using software that gave below acceptable
results. This patch adds the necessary to use the Mediacodec encoders
recently available in ffmpeg.
For that to happen it was necessary to change the image reader of the
QtCamera2 to use YUV420p frame instead of JPEG. That also improves
performance since decoding is no longer necessary. This pipeline only
has one copy done on FFMpeg side to upload to Mediacodec input buffer.

Fixes: QTBUG-102235
Pick-to: 6.5 6.5.0
Change-Id: Ief5421c700dd0a77763a7713f9b452dd01828241
Reviewed-by: Lars Knoll <lars@knoll.priv.no>