Improve ffmpeg hw accel and codecs finding
A bunch of the improvements mostly related to linux.
The most problem with linux hw acceleration is that there're no exact
working acceleration for all hardwares.

On linux we had only hw acceleration 'vaapi' that requires
installation of vaapi drivers, also on some hardwares
it's not supported or not stable.
Pretty often ffmpeg returns vaapi as supported hw device but
it failes on creation vaapi hw context.
The same might happen with other accelerations as well.

The commit makes selection and creation of hw devices
with codec for them more reliable and flexible,
also, it extends the list of possible hw accel devices for linux,
and makes customization of codecs choosing easier in the future.

The functionality has been checked with cuda and vdpau hw accelerations
on PC with nvidia rtx a2000. Both decoding and encoding work fine,
handling of vdpau textures is to be done in the future.
The hope and prospective is that will work on
embedded platforms with nvidia gpu, it is to be tested afterwards.

Since hw accel types are very platform specific thing,
I propose runtime env vars
QT_FFMPEG_ENCODING_HW_DEVICE_TYPES and QT_FFMPEG_DECODING_HW_DEVICE_TYPES
to allow advanced users configuring priorities of used acceleration for
the exact target platform.

The changes have been tested on linux, macos, windows.

Should be done afterwards:
- conversion vdpau textures to opengl, QTBUG-111593
- add vaapi, vdpau, cuda to linux CI builds. I've checked that their
  adding don't affect ffmpeg lib size to much, less than 1%.
- describe hw acceleration details and some setup guide
  in the qt documentation.
- handle codec open fail during codecs search.

Pick-to: 6.5 6.5.0
Change-Id: I8f120f53573f928a0673ed4ad94607a0ac57822d
Reviewed-by: Lars Knoll <lars@knoll.priv.no>