Android: Fix stretched camera view in portrait mode
The camera view in portrait mode was displayed as landscape.

After 579a0423b8b3c62c582a706d338a9eb1bcd788a3 commit, the resolution is
not transposed if it is not in supported preview sizes. That is the
expected way for camera preview size, but video output resolution need
to be transposed anyway.

This commit rotates the resolution of the video output in portrait mode,
even if the camera resolution is not rotated due to unsupported size.

Fixes: QTBUG-110975
Fixes: QTBUG-111266
Fixes: QTBUG-111421
Fixes: QTBUG-107173
Pick-to: 6.2 6.4 6.5
Change-Id: I773fb8d0df1b616c6af8db9a8d7ad785da8295ee
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>