Detect QVideoFrame changes in QSGVideoNode
When recalculating the texture geometry of a QSGVideoNode the parameters are
checked to avoid costly recalculation. But the check does not include the
QVideoFrame orientation and mirrored state. Therefor changes in these properties
are not considered and don't trigger an update of the geometry.

Change-Id: I0eeab583865bbc03a5b75a948063d2e13d12fe7a
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 40740c4b46731de1c56166f206230a9219814957)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>