Android: Postpone FFmpeg camera activation
Activation of the Android FFmpeg camera should be postponed until the
first frame is read. Handling the first frame allows the implementation
to get information about the pixel format of the frame and prepare the
encoder. Until this happnes, we are unable to use the camera effectively
(e.g. for recording).

Task-number: QTBUG-99098
Change-Id: Id0574b25318cb8cb09c70cca12e21f8970e31dac
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 75b9724ed81393492460c255937c2c633072878e)
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>