Temporary disable QFFmpegScreenCaptureUwp
The windows implementation QFFmpegScreenCaptureUwp is not stable,
on Windows 11 it works, on Windows 10 it throws asserts, crashes,
or produces side effects.
The other implementation, QFFmpegScreenCaptureDxgi, works better.

The suggestion is to turn off QFFmpegScreenCaptureUwp in the release
so that QFFmpegScreenCaptureDxgi can be used instead, and we provide
working functionality.

Task-number: QTBUG-108892
Pick-to: 6.5 6.5.0
Change-Id: I027194572ae141ec154c566f88371a1aa8adc7de
Reviewed-by: Lars Knoll <lars@knoll.priv.no>