Add specific env var for testing without hw textures conversion
The reason of this is that textures conversion may be is not stable
on the user's side, and we need an opportunity to check if this
problem affects the user's issue. For now, we already have
such a problem to be checked.

Task-number: QTBUG-111459
Change-Id: I7f432fb7fcd0c6e1dc8b9b1b8e6d100d966a4076
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 24cfd4a28a06eb00804265ec87f54e6f5d099fd0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>