Filter codecs if hw acceleration is not available
The main reason of the patch is the trickiness of hw acceleration
customization on Linux. In some cases hw accelerated ffmpeg encoders
doesn't work even though they are declared as available,
and the user is able to turn off the encoding vaapi hw acceleration.

In this case, on the initialization of QFFmpegMediaFormatInfo,
the function findAVEncoder shouldn't consider codecs
with disabled hw accel as available.

Pick-to: 6.5
Change-Id: Iafe2a9b1bb5e635882a974f752fe0311539223b2
Reviewed-by: Lars Knoll <lars@knoll.priv.no>