Tests: fix failing tst_qaudiodecoderbackend on Android
Test set contains testFileUrl method. Inside the method QTemporaryFile
was used for Android. Calling deleteLater on it, schedules this object
for deletion. On some tests (like restartOnBufferReady) temporaryFile
was destroyed too fast and the test may be finished with error because
of that. 

This commit contains the fix for a better control of deletion
QTemporaryFile for Android (in tst_qaudiodecoderbackend).

Fixes: QTBUG-112171
Change-Id: I568c75374aa16018b5dc67d31b733ceb65e91a70
Reviewed-by: Lars Knoll <lars@knoll.priv.no>