Tests: fix for tst_QAudioSink testset part 1/2
tst_QAudioSink testset is currently skipped on CI. Need to be fixed.
This is part of the fix related with running tests on Android.

SampleRate was set to 1024 in pullResumeFromUnderrun test. It seems that
the sampleRate was too low and may not be supported by some Android
Devices. According OpenSLES implementation, the lowest hardcoded Sample
Rate is 8000000 milliHertz (8000Hz)[0].

This commit changed sampleRate in pullResumeFromUnderrun test from 1024
to 8000.

[0]https://chromium.googlesource.com/android_ndk.git/+/refs/tags/r14b/sysroot/usr/include/SLES/OpenSLES.h#157

Task-number: QTBUG-112172
Change-Id: Iebc673558afb1219423ccd138539284e42d9c8be
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit aff461fbafa11f111819e2c830d851524d3825ab)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>