Tests: fix for tst_QAudioSink testset part 2/2
tst_QAudioSink testset is currently skipped on CI. Need to be fixed.
This is part of the fix related with running tests on Android.

In the pullResumeFromUnderrun test, the signalEnd was set to true before
resuming pull. That is why, AudioPullSource::atEnd will also return true
after reading all the data.
UnderrunError should be set when "Audio data is not being fed to the
audio device at a fast enough rate", but the error should not be set
when all data has already been read. That is why it should not be
expected to get UnderrunError at the end of the test.

This commit removes comparing error state to UnderrunError at the end of
the pullResumeFromUnderrun test.

Task-number: QTBUG-112172
Pick-to: 6.5
Change-Id: Ia2a0e40c3227a7819d1bf275f04738a0e780eee0
Reviewed-by: Lars Knoll <lars@knoll.priv.no>