Android: fix readyRead method in qandroidaudiosink
In the pull mode, in case when new data is ready to be read from
QIODevice, the readyRead() method is used. In such case we need to call
Enqueue() from OpenSL ES. After calling it, OpenSL ES will start again
using bufferQueueCallback for his own, when the data will be needed.

Before this commit, Enqueue() was called from the same thread which
readyRead signal comes from. That causes the state changes overlapping.
InvokeMethod is now used to solve the problem.

Task-number: QTBUG-112172
Pick-to: 6.5
Change-Id: I765cff90f0bf7595e28c0cd8171b0bcbfda8d392
Reviewed-by: Lars Knoll <lars@knoll.priv.no>