AVFMediaPlayer: Try to play asset even if 'playable' property is false
Sometimes an AVAsset can report to be not playable even if the asset
can still be played. That is why we shouldn't stop trying to play based
on this check only. To determine if an asset is truly playable or not,
we should only need to rely on the 'status' property of the AVPlayerItem
using the asset, as that one will never be AVPlayerStatusReadyToPlay
if the asset is not playable and will instead fail throwing the
appropriate errors.

Fixes: QTBUG-98102
Pick-to: 6.2 6.5
Change-Id: I26d319fe120335cdfbc09cd6aa68ce7e8f282699
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>