Tests: fix failing tst_QCameraBackend on Android
This is a commit already delivered to dev, but missed cherry-pick to 6.5

On Android, camera activation take some time. It may happen that we are
trying to start recording on not active camera - which cause test
failure. That is why we need to wait for camera activation before test
continuation.

This commit extends the waiting time for an activity state change.

Task-number: QTBUG-112175
Change-Id: Ife9d03ff20fa9136a86fd743a63fb9653696c59f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>