Tests: fix failing tst_QCameraBackend on Android part 3/3
YUV420P and NV12 are similar formats. They contain the same data, just
slightly differently arranged in memory.
On Android we cannot be sure which format we will receive. It dependes
on vendors' implementaion. Therefore, it is not a failure when we asked
for YUV420P but got NV12.

This commit contains fix in comparation format types for Android.
Formta_YUV420P and Fromat_NV12 are treated as the same formats.

Task-number: QTBUG-112175
Change-Id: Ie166226f43a42b55cfe18864594ff1111cbf79ef
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>