AVFCameraRenderer: Properly initialize output settings
Since we are allocating and initializing m_outputSettings for the
video output, we need to do the same for the camera output
(AVFCameraRenderer), as the settings are then released for both
in the base class (AVFVideoSinkInterface) destructor. Releasing these
without proper initialization in AVFCameraRenderer was causing a crash.
Ammends 2e43d29e1d5d50b44b8f6d4f000968e3933c279a.

Fixes: QTBUG-110131
Fixes: QTBUG-112352
Pick-to: 6.5 6.2
Change-Id: Ibb0a42bb2dc75522ede24f479a13be92599801c5
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>