AVFCamera: Disconnect the recorder receiver when changing cameras
We are connecting the darwin recoder (AVFMediaEncoder) to the
camera's activeChanged signal in order to properly stop recordings
when camera is deactivated. However, if the camera changes, we need to
disconnect the recorder from the old camera's signals.

Change-Id: I5fb566ca52cdd558d8f33d4df8650e0734d33292
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 32d3a78677ca225f7037f53b1ce291490ad29686)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>