ffmpeg: Clear the last frame in videoSink after stopping the player
The frame buffer was not properly cleared after stopping video. The
last frame was still kept by QPlatformVideoSink (even if video was
stopped). That caused the issue when trying to configure the
MediaCodec for the next video file.

This commit contains clearing the last frame in videoSink, after
stopping video on the qffmpegmediaplayer.

Task-number: QTBUG-112173
Pick-to: 6.5
Change-Id: I0d5a93666db2dbac7577ef5af994e34141192105
Reviewed-by: Lars Knoll <lars@knoll.priv.no>