Android-tests: Fix tst_qscreencapture_integration part 2
tst_QScreenCaptureIntegration testset is currently Blacklisted on CI for
Android platform (also for Windows, MacOs and Linux). This commit
contains part of fix for the testset.

Expected captured screen size was calculated as multiplication screen
size and device pixel ratio. Such multiplication may not return an
integer value (when device pixel ratio is not integer). For example when
screen size is 393x873 and pixelRatio is 2,75 expected caprutered screen
size is 1081x2401 (because of rounding 1080,75x2400,75).

This commit, to avoid wrong rounded values, uses qFloor to calculate the
expected size of the captured screen.

Task-number: QTBUG-112176
Change-Id: Ib3daaf93e609b96e2d51efb8a4640f45a9125329
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>