Android-tests: Fix tst_qscreencapture_integration part 5
tst_QScreenCaptureIntegration testset is currently Blacklisted on CI for
Android platform (also for Windows, MacOs and Linux). This commit
contains part of fix for the testset.

This commit turns on mentioned testset for Android on CI. Only
recordToFile test will be still blacklisted according to QTBUG-112827

Task-number: QTBUG-112176
Task-number: QTBUG-112827
Change-Id: I6b9a35d663bf980550a7f3d91b172de007ede5d3
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>