Fix playing of video files with Chinese names
The old implementation was using QUrl::toEncoded in order to retrieve a
filename from a URL, but it uses toLatin1 internally which substitutes
all non-ASCII characters with question marks that leads to ill formed
path. The new implementation uses QUrl::toString with a subsequent call
to QString::toUtf8 and it basically leads to the same results while
preserving non-ASCII characters.

Task-number: QTBUG-111951
Pick-to: 6.5
Change-Id: I0ad32143bc4afbdaec34511d7910ff106cb3cf45
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>