Fix crash with sequential QIODevice as media source on Windows
The sample in QTBUG-110005 uses a QNetworkReply as source for
QMediaPlayer. This is a sequential QIODevice. As such, it doesn't have
the concept of size.

The sequence leading to the crash is
- BeginCreateObjectFromByteStream() is called
- MFStream::GetLength() return zero
- that's why MFStream::BeginRead() is not called,
  and m_currentReadResult stays nullptr
- QNetworkReply::readyRead() is emitted
- MFStream::doRead() is called, because of the latter signal
- MFStream::doRead() accesses m_currentReadResult,
  which is still nullptr (see above), and we witness the crash

The desirable sequence is (as one can inspect by playing a video from a
qrc:// URL):
- BeginCreateObjectFromByteStream() is called
- MFStream::GetLength() returns the video stream's size
- MFStream::BeginRead() is called and initializes m_currentReadResult
- MFStream::BeginRead() posts a custom event to the MFStream object
- MFStream::doRead() is called when handling that custom event
- MFStream::EndRead() tears down m_currentReadResult

There is never a situation where QIODevice::readRead() should trigger a
MFStream::doRead(). Calling doRead() outside of the
BeginRead()/EndRead() pair cannot work. Therefore, this patch removes
the handling of the stream's readyRead() signal.

Note that this patch only fixes the crash in favor of an error code
that's written to the program's console. Handling sequential QIODevice
streams will require some more work. See further comments in the issue.

Fixes: QTBUG-110005
Change-Id: I2fc70f873a2e3a257e8e081d83862c53dabb6cac
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit d7c1429b5515c38288e3a33992bf2582c243df5a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>