Android-backend: Fix not working camera preview
In case when the QML CaptureSession was opened from QuickWidget, it can
be be started before Rhi was set in videoSink. In such case,
surfaceTexture without Rhi will be used for setPreviewTexture.
It may happen that Rhi will be set later, but CameraSession will not be
informed about it.

For this case, rhiChanged Signal is moved to QPlatformVideoSink.
This allows to restart cameraSession when Rhi is set later.

Pick-to: 6.6 6.5
Fixes: QTBUG-112226
Change-Id: If95fd7f390664e0b999e904a97e0d48c5b4b6c7c
Reviewed-by: Lars Knoll <lars@knoll.priv.no>