QMediaPlayer: emit errorChanged signal whenever the error has changed
There was a problem that error/errorString properties were not properly
notified when reseted on start playing. So now we use private setError()
method whenever 'error' should be changed.

Also as a drive-by there are several improvements:
- errorChanged signal will be emitted now only when the property's
value was actually changed.
- errorOccured signal is now emitted only when there is a real error,
i.e. it will never be emitted with NoError.

Fixes: QTBUG-113386
Pick-to: 6.5
Change-Id: I16e26710941ed6b8429f39c201572092f2714c74
Reviewed-by: Lars Knoll <lars@knoll.priv.no>