Android-backend: Fix for Captured Image rotation
When using android-camera backend, the result photo was not correctly
rotated for landscape orientation.

Preview rotation is handled by [0]setDisplayOrientation method, but it
does not affect taking pictures and recording. Recording is handled by
[1]setOrientationHint.

Both mentioned methods use QAndroidCameraSession::currentCameraRotation
for rotate correction. It is reasonable to handle photo rotation in the
same way.

This commit removes a special rotation handling from onPictureTaken java
method and move it to QAndroidCameraSession. That allows to handle photo
rotation in the same way as preview rotation and recording rotation.

https://developer.android.com/reference/android/hardware/Camera#setDisplayOrientation(int)
https://developer.android.com/reference/android/media/MediaRecorder#setOrientationHint(int)

Fixes: QTBUG-113029
Pick-to: 6.5 6.2
Change-Id: I2d08fef43fbd78faeafc7e20e100329e75019679
Reviewed-by: Lars Knoll <lars@knoll.priv.no>