Android-backend: fix for setting maximumZoomFactor
Before this commit, the maximum zoom factor was set locally in the
QAndroidCamera class. It was not propagated to its base QPlatformCamera
class. That cause the problem with wrong value returned by
maxZoomFactor() method.

This commit contains setting correctly m_maxZoom in QPlatformCamera and
remove not needed m_maximumZoom member from QAndroidCamera.

Fixes: QTBUG-105372
Pick-to: 6.5 6.2
Change-Id: Iaf167d03e42cb0e22ebe60862bb9dc24b1ef9173
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>