Improve naming of screen capture integration tests
Reviewing test logs and maintaining tests can be easier if the tests
are explicit in what they test. It can therefore be useful to document
in the tests what function is being tested, what the expectation is, and
which conditions applies to a test.

The proposed naming turns the tests into requirements, and can be read
as (Require that) FunctionName DoesWhat UnderWhichConditions. For
example:

    setWindow_selectsWindow_whenWindowIsOverlapped

Here, the function under test is setWindow. When it is called, we expect
that the window is selected (when capturing). The condition is that the
window is overlapped.

In some cases, the test may not test a single function, where the
proposal is to do as best as we can to document what functionality is
tested, what the expectation is, and under what conditions the test is
performed.

Task-number: QTBUG-113460
Change-Id: Ie001e464a4c92d468b5b58160b1cdaa2aea858ee
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 2d65ac72a38ca8e280758db7430756dc1a468628)