Commonize popup behavior between camera settings and device list
Currently when you select a device in the in the device list popup
it does not close the popup, unlike with the camera settings popup list
This leads to some user-confusion regarding expected behavior of the
example.

I replicated the emission of the selected() signal as used in the
camera settings popup to fix this.

Pick-to: 6.5 6.2
Change-Id: I1ef5faf75dff702f61991b9ac1735364b187b10b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>