Limit VideoEncoder memory usage when encoder is unable to keep up
The qffmpegencoder has an internal queue that buffers video frames when
the encoder is unable to keep up with the frame source. This helps
stabilizing the frame rate in case of intermittent encoder hiccups, but
if the encoder is consistently unable to keep up with the frame source,
the queue will grow and eventually the process will run out of memory.

With 4K resolution images, the data rate is approx 2 GB/second, and if
the frames are located on the GPU, the GPU will run out of memory in a
matter of seconds.

To prevent this, we introduce a max size of the VideoEncoder frame
queue. In this change proposal, we use an arbitrarily chosen limit of 10
frames. This should be enough to recover from short intermittent encoder
hiccups, without using too much memory (~300 MB at 4K resolution).

If a new frame arrives from the video source, and the queue is full, we
now drop the frame. This strategy prevents blocking any video source
thread (as opposed to blocking if the queue is full.)

Fixes: QTBUG-114055
Pick-to: 6.5 6.6
Change-Id: I29714267696454dfd371e0fd6ed3112c9b7356b0
Reviewed-by: Lars Knoll <lars@knoll.priv.no>