Align QComPtr constructor ownership with other COM smart pointers
COM smart pointers such as CComPtr, _com_ptr_t and WRL::ComPtr all calls
AddRef when constructed with a valid interface pointer. To make the
QComPtr less 'special', it should also call AddRef when assigned a valid
pointer.

This is a step towards replacing QComPtr out with WRL::ComPtr which is used elsewhere in Qt.

Pick-to: 6.5 6.6
Change-Id: I97574c28267537f11b6d73c28da9ccf54a6cdcf1
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>