Split QComPtr reset() function into a reset() and an attach() function
The name 'reset' does not specify well how ownership is handled when
reset to a new pointer value. A better name for this is 'attach', which
indicates that AddRef is not called.

'reset' now just resets the pointer to a default constructed state.

This is a step towards replacing QComPtr out with WRL::ComPtr which is used elsewhere in Qt.

Pick-to: 6.5 6.6
Change-Id: Ic72beba8cf212669f5dfa2e01f36754ffb0ad94e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>