Introduce a QComPtr friendly factory function for COM objects
In QtMultimedia, some COM objects are constructed with `new`, and they
are constructed with a reference count of 1. By introducing a factory
function for COM objects, we avoid the need to think too much when
creating COM objects without using CoCreateInstance.

This pattern is in line with Windows Runtime Library's WRL::Make and
WinRT's winrt::make_self helper functions.

Pick-to: 6.5 6.6
Change-Id: Ife134fed1a85a614eedd9d8a17d53857d10fcedf
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>