Add CMakeUserPresets.json to .gitignore in QtMultimedia
The CMakeUserPresets.json file can be used to simplify configuring Qt,
particularly with tools such as Visual Studio and Visual Studio Code.

As opposed to the CMakePresets.json, the CMakeUserPresets.json should
only contain user specific configurations, and should never be checked
into source code management system. This was added to the top level
repo in but to ease working with QtMultimedia as a module, it should
also be included here.

Pick-to: 6.5 6.6
Change-Id: I003017c6049860fa306a2353aab00381378878d7
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>